move codes from gdp_test to google_discovery_protocol_packetin_test and remove argument t in gdp_test function and related interfaces
